Revert "Add service to set ostree bootloader on existing systems"
This reverts commit 753e3e7ef3097eb61f5335091fedc616310d0efe. As its
message says:

> After a checkpoint we can drop this script and unit.

This commit was first included in 5.0.0 but it was backported to 4.0.10
as 6c9ea9ac4a886fbb63bbcccf43d783e8c8d8fd97. So all systems running
Endless OS 6 have passed through 2 checkpoints; and the eos4.0
version of eos-updater checks that this configuration field is set
before allowing the system to update to 5.x.