Add eos-update-efi-uuid for updating EFI load option partition UUIDs
When the partition UUID for the ESP is updated, any load options
referencing the partition become invalid. Add a tool to update all load
options using the old partition UUID to the new partition UUID. This
uses the libefivar and libefiboot packages to access EFI variables and
interpret the data appropriately.

https://phabricator.endlessm.com/T31604