Revert "eos-test-mode: Remount eos-live overlays in some services' namespace"
This reverts commit 28e58e8f7a8c0e83de079014915dbbcca7cdec30.

The underlying bug in ostree was fixed in
https://github.com/endlessm/ostree/pull/214, a backport of
https://github.com/ostreedev/ostree/pull/3292.

With that change, eos-test-mode's overmounted overlayfses propagate
correctly to NetworkManager and AccountsService, so they no longer need
to be restarted. And restarting NetworkManager triggers a bug in Initial
Setup, which does not correctly handle this case and so never lists
available Wi-Fi networks.

https://phabricator.endlessm.com/T35640