test-mode: Don't pass etc as extra path to live-boot-overlayfs-setup
I believe this was needed in the distant path where /etc was handled
differently to other mutable paths on live boots, but since commit
4d348fb9b56ff34cf150d68a6012b6ce5778d51a and commit
37d20f1512958ed9ec3f6d7d776c760cbcb6b882 etc has been part of the
overlay_dirs list that EXTRA_PATHS is appended to, so passing etc to
this script has been redundant. Remove it; and remove the EXTRA_PATHS
handling from eos-live-boot-overlayfs-setup, which is now unused.

https://phabricator.endlessm.com/T35651