overlayfs-setup: Fallback to read only mount when overlay not supported
Overlayfs doesn't support using case insensitive filesystems such as
vfat and exfat as a lower directory. Change the mount to read only so
changes can't be made there.

In theory this would simply be `mount -o remount,ro`. Unfortunately,
if systemd has setup an automount, autofs is not our friend. In that
case, the mount options need to be persisted into the mount unit so that
a subsequent remounting retains the ro option.

https://phabricator.endlessm.com/T35641