Partially revert "Remove obsolete cleanup of obsolete stamp files"
This reverts part of commit 1338aae7721a5bfddc8b121fb1b761bd779d4313.

Sadly I had not realised that this stamp file was also created by a hook
in the image builder. So on every new system installed between 3.5.0 and
6.0.1 inclusive, this stamp file was preinstalled and deleted on first
boot; while on all systems installed with 6.0.2 or newer, the stamp file
was preinstalled but never removed.

https://github.com/endlessm/eos-image-builder/commit/6d1389f4043941593d9c47060a70753a1bdd2e65
removes the image builder hook, but for the sake of those systems
installed with 6.0.2 or newer, reinstate the tmpfiles snippet that
deletes the stamp file.